pyproject.toml
[project]
name = "integrate-baumer-cam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "baumer-neoapi",
    "numpy>=2.2.6",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "pymodbus>=3.7.0",
    "requests>=2.32.5",
]

[tool.uv.sources]
baumer-neoapi = { path = "libs/baumer_neoapi-1.5.0-cp34.cp35.cp36.cp37.cp38.cp39.cp310.cp311.cp312-none-win_amd64.whl" }
